+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # 更新日志
 
+### 1.0.1
+##### honoka-utils 1.0.1
+- 将原属于qqrobot-spring-boot-starter的`ImageUtils`类移到本库中。
+
 ### 1.0.0
 ##### honoka-json-fastjson 1.0.0
 - 起始版本